requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
opencv-python
pytesseract
ultralytics
numpy
python-jose[cryptography]
pydantic